requirements.txt
aiohttp>=3.9.5
aiorpcX>=0.23.0
coloredlogs>=15.0.1
base58>=2.1.1
pyzmq>=25.1.0
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
aiosqlite>=0.19.0

